Fix benchmark duplication.
Previously we were doing benchmark discovery from the YAML file twice.
We loaded the benchmarks.yml file when initializing a new RubyBench
object, this then cached the benchmark names and types in instance
variables in the RubyBench object.

Then in the main loop we were loading the yaml again, and feeding each
benchmark one at a time to the main RubyBench.run_benchmark function in
a loop.

This resulted in the main benchmark loop attempting to run the
ractor-only benchmarks with the main run_benchmark loop using the
non-ractor harness. These would fail, and record nil results in a set of
files in the `ruby-bench/ractor` directory in the rubybench-data repo.

This commit extracts the logic to read the yaml and categorize the
benchmarks into a single object.

After this change RubyBench is responsible for controlling ruby-bench to
run a named benchmark inside a container.

RubyBench::Benchmarks is responsible for categorising the benchmarks
from the YAML.

eg.

b = RubyBench::Benchmarks.parse_benchmark_file("benchmarks.yml")

b.regular = [] # Benchmarks to run with any harness

b.ractor_compatible = [] # Benchmarks that can be run with the ractor
                         # harness

b.ractor_only = [] # Benchmarks that can _only_ be run with the ractor
                   # harness